requirements/dev.txt
-r base.txt
bpython==0.12
django-debug-toolbar==0.9.4
